pyproject.toml
[project]
name = "sql-nosql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg>=3.3.3",
    "psycopg-pool>=3.3.0",
    "python-dotenv>=1.2.2",
]
